Fix benchmarking and move helper functions to testing_utils
