+++ pallets/redeem/README.md
@@ -1,0 +1,36 @@
+# Issue pallet
+
+## Testing
+
+To run the tests use:
+
+```bash
+cargo test --package redeem --features runtime-benchmarks
+```
+
+## Benchmarking
+
+Build the node with the `runtime-benchmarks` feature:
+
+```bash
+cargo build --package spacewalk-standalone --release --features runtime-benchmarks
+```
+
+```bash
+# Show benchmarks for this pallet
+./target/release/spacewalk-standalone benchmark pallet -p redeem -e '*' --list
+```
+
+Run the benchmarking for a pallet:
+
+```bash
+./target/release/spacewalk-standalone benchmark pallet \
+--chain=dev \
+--pallet=redeem \
+--extrinsic='*' \
+--steps=100 \
+--repeat=10 \
+--wasm-execution=compiled \
+--output=pallets/redeem/src/default_weights.rs \
+--template=./.maintain/frame-weight-template.hbs
+```
