+++ pallets/redeem/src/benchmarking.rs
@@ -1,29 +1,27 @@
-use super::*;
-use bitcoin::{
-	formatter::{Formattable, TryFormattable},
-	types::{
-		BlockBuilder, RawBlockHeader, TransactionBuilder, TransactionInputBuilder,
-		TransactionInputSource, TransactionOutput,
-	},
-};
-use btc_relay::{BtcAddress, BtcPublicKey};
-use currency::getters::{get_relay_chain_currency_id as get_collateral_currency_id, *};
 use frame_benchmarking::{account, benchmarks, impl_benchmark_test_suite};
 use frame_support::assert_ok;
 use frame_system::RawOrigin;
 use orml_traits::MultiCurrency;
-use primitives::{CurrencyId, CurrencyId::Token, TokenSymbol::*, VaultCurrencyPair, VaultId};
 use sp_core::{H256, U256};
 use sp_runtime::traits::One;
 use sp_std::prelude::*;
-use vault_registry::types::Vault;
+
+use currency::getters::{get_relay_chain_currency_id as get_collateral_currency_id, *};
+use oracle::Pallet as Oracle;
+use primitives::{CurrencyId, CurrencyId::Token, TokenSymbol::*, VaultCurrencyPair, VaultId};
+use security::Pallet as Security;
+use stellar_relay::{
+	testing_utils::{
+		build_dummy_proof_for, get_validators_and_organizations, DEFAULT_STELLAR_PUBLIC_KEY,
+	},
+	Pallet as StellarRelay,
+};
+use vault_registry::{types::Vault, Pallet as VaultRegistry};
 
 // Pallets
 use crate::Pallet as Redeem;
-use btc_relay::Pallet as BtcRelay;
-use oracle::Pallet as Oracle;
-use security::Pallet as Security;
-use vault_registry::Pallet as VaultRegistry;
+
+use super::*;
 
 type UnsignedFixedPoint<T> = <T as currency::Config>::UnsignedFixedPoint;
 
@@ -37,7 +35,7 @@ fn wrapped<T: crate::Config>(amount: u32) -> Amount<T> {
 
 fn register_public_key<T: crate::Config>(vault_id: DefaultVaultId<T>) {
 	let origin = RawOrigin::Signed(vault_id.account_id.clone());
-	assert_ok!(VaultRegistry::<T>::register_public_key(origin.into(), BtcPublicKey::dummy()));
+	assert_ok!(VaultRegistry::<T>::register_public_key(origin.into(), DEFAULT_STELLAR_PUBLIC_KEY));
 }
 
 fn deposit_tokens<T: crate::Config>(
@@ -77,70 +75,6 @@ fn initialize_oracle<T: crate::Config>() {
 	Oracle::<T>::begin_block(0u32.into());
 }
 
-fn mine_blocks_until_expiry<T: crate::Config>(request: &DefaultRedeemRequest<T>) {
-	let period = Redeem::<T>::redeem_period().max(request.period);
-	let expiry_height = BtcRelay::<T>::bitcoin_expiry_height(request.btc_height, period).unwrap();
-	mine_blocks::<T>(expiry_height + 100);
-}
-
-fn mine_blocks<T: crate::Config>(end_height: u32) {
-	let relayer_id: T::AccountId = account("Relayer", 0, 0);
-	mint_collateral::<T>(&relayer_id, (1u32 << 31).into());
-
-	let height = 0;
-	let block = BlockBuilder::new()
-		.with_version(4)
-		.with_coinbase(&BtcAddress::dummy(), 50, 3)
-		.with_timestamp(1588813835)
-		.mine(U256::from(2).pow(254.into()))
-		.unwrap();
-
-	let raw_block_header = RawBlockHeader::from_bytes(&block.header.try_format().unwrap()).unwrap();
-	let block_header = BtcRelay::<T>::parse_raw_block_header(&raw_block_header).unwrap();
-
-	Security::<T>::set_active_block_number(1u32.into());
-	BtcRelay::<T>::_initialize(relayer_id.clone(), block_header, height).unwrap();
-
-	let transaction = TransactionBuilder::new()
-		.with_version(2)
-		.add_input(
-			TransactionInputBuilder::new()
-				.with_source(TransactionInputSource::FromOutput(block.transactions[0].hash(), 0))
-				.with_script(&[
-					0, 71, 48, 68, 2, 32, 91, 128, 41, 150, 96, 53, 187, 63, 230, 129, 53, 234,
-					210, 186, 21, 187, 98, 38, 255, 112, 30, 27, 228, 29, 132, 140, 155, 62, 123,
-					216, 232, 168, 2, 32, 72, 126, 179, 207, 142, 8, 99, 8, 32, 78, 244, 166, 106,
-					160, 207, 227, 61, 210, 172, 234, 234, 93, 59, 159, 79, 12, 194, 240, 212, 3,
-					120, 50, 1, 71, 81, 33, 3, 113, 209, 131, 177, 9, 29, 242, 229, 15, 217, 247,
-					165, 78, 111, 80, 79, 50, 200, 117, 80, 30, 233, 210, 167, 133, 175, 62, 253,
-					134, 127, 212, 51, 33, 2, 128, 200, 184, 235, 148, 25, 43, 34, 28, 173, 55, 54,
-					189, 164, 187, 243, 243, 152, 7, 84, 210, 85, 156, 238, 77, 97, 188, 240, 162,
-					197, 105, 62, 82, 174,
-				])
-				.build(),
-		)
-		.build();
-
-	let mut prev_hash = block.header.hash;
-	for _ in 0..end_height {
-		let block = BlockBuilder::new()
-			.with_previous_hash(prev_hash)
-			.with_version(4)
-			.with_coinbase(&BtcAddress::dummy(), 50, 3)
-			.with_timestamp(1588813835)
-			.add_transaction(transaction.clone())
-			.mine(U256::from(2).pow(254.into()))
-			.unwrap();
-		prev_hash = block.header.hash;
-
-		let raw_block_header =
-			RawBlockHeader::from_bytes(&block.header.try_format().unwrap()).unwrap();
-		let block_header = BtcRelay::<T>::parse_raw_block_header(&raw_block_header).unwrap();
-
-		BtcRelay::<T>::_store_block_header(&relayer_id, block_header).unwrap();
-	}
-}
-
 fn test_request<T: crate::Config>(vault_id: &DefaultVaultId<T>) -> DefaultRedeemRequest<T> {
 	RedeemRequest {
 		vault: vault_id.clone(),
@@ -148,11 +82,11 @@ fn test_request<T: crate::Config>(vault_id: &DefaultVaultId<T>) -> DefaultRedeem
 		period: Default::default(),
 		fee: Default::default(),
 		transfer_fee: Default::default(),
-		amount_btc: Default::default(),
+		amount: Default::default(),
+		asset: Token(DOT),
 		premium: Default::default(),
 		redeemer: account("Redeemer", 0, 0),
-		btc_address: Default::default(),
-		btc_height: Default::default(),
+		stellar_address: Default::default(),
 		status: Default::default(),
 	}
 }
@@ -170,7 +104,8 @@ benchmarks! {
 		let origin: T::AccountId = account("Origin", 0, 0);
 		let vault_id = get_vault_id::<T>();
 		let amount = Redeem::<T>::redeem_btc_dust_value() + 1000u32.into();
-		let btc_address = BtcAddress::dummy();
+		let asset = Token(DOT);
+		let stellar_address = DEFAULT_STELLAR_PUBLIC_KEY;
 
 		initialize_oracle::<T>();
 
@@ -192,7 +127,7 @@ benchmarks! {
 		assert_ok!(Oracle::<T>::_set_exchange_rate(get_collateral_currency_id::<T>(),
 			UnsignedFixedPoint::<T>::one()
 		));
-	}: _(RawOrigin::Signed(origin), amount, btc_address, vault_id.clone())
+	}: _(RawOrigin::Signed(origin), amount, asset, stellar_address, vault_id.clone())
 
 	liquidation_redeem {
 		assert_ok!(Oracle::<T>::_set_exchange_rate(get_collateral_currency_id::<T>(),
@@ -232,11 +167,11 @@ benchmarks! {
 
 		initialize_oracle::<T>();
 
-		let origin_btc_address = BtcAddress::dummy();
+		let origin_stellar_address = DEFAULT_STELLAR_PUBLIC_KEY;
 
 		let redeem_id = H256::zero();
 		let mut redeem_request = test_request::<T>(&vault_id);
-		redeem_request.btc_address = origin_btc_address;
+		redeem_request.stellar_address = origin_stellar_address;
 		Redeem::<T>::insert_redeem_request(&redeem_id, &redeem_request);
 
 		register_public_key::<T>(vault_id.clone());
@@ -251,66 +186,16 @@ benchmarks! {
 			vault
 		);
 
-		let height = 0;
-		let block = BlockBuilder::new()
-			.with_version(4)
-			.with_coinbase(&origin_btc_address, 50, 3)
-			.with_timestamp(1588813835)
-			.mine(U256::from(2).pow(254.into())).unwrap();
-
-		let block_hash = block.header.hash;
-		let raw_block_header = RawBlockHeader::from_bytes(&block.header.try_format().unwrap()).unwrap();
-		let block_header = BtcRelay::<T>::parse_raw_block_header(&raw_block_header).unwrap();
-
 		Security::<T>::set_active_block_number(1u32.into());
-		BtcRelay::<T>::_initialize(relayer_id.clone(), block_header, height).unwrap();
 
-		let value = 0;
-		let transaction = TransactionBuilder::new()
-			.with_version(2)
-			.add_input(
-				TransactionInputBuilder::new()
-					.with_source(TransactionInputSource::FromOutput(block.transactions[0].hash(), 0))
-					.with_script(&[
-						0, 71, 48, 68, 2, 32, 91, 128, 41, 150, 96, 53, 187, 63, 230, 129, 53, 234,
-						210, 186, 21, 187, 98, 38, 255, 112, 30, 27, 228, 29, 132, 140, 155, 62, 123,
-						216, 232, 168, 2, 32, 72, 126, 179, 207, 142, 8, 99, 8, 32, 78, 244, 166, 106,
-						160, 207, 227, 61, 210, 172, 234, 234, 93, 59, 159, 79, 12, 194, 240, 212, 3,
-						120, 50, 1, 71, 81, 33, 3, 113, 209, 131, 177, 9, 29, 242, 229, 15, 217, 247,
-						165, 78, 111, 80, 79, 50, 200, 117, 80, 30, 233, 210, 167, 133, 175, 62, 253,
-						134, 127, 212, 51, 33, 2, 128, 200, 184, 235, 148, 25, 43, 34, 28, 173, 55, 54,
-						189, 164, 187, 243, 243, 152, 7, 84, 210, 85, 156, 238, 77, 97, 188, 240, 162,
-						197, 105, 62, 82, 174,
-					])
-					.build(),
-			)
-			.add_output(TransactionOutput::payment(value.into(), &origin_btc_address))
-			.add_output(TransactionOutput::op_return(0, H256::zero().as_bytes()))
-			.build();
-
-		let block = BlockBuilder::new()
-			.with_previous_hash(block_hash)
-			.with_version(4)
-			.with_coinbase(&origin_btc_address, 50, 3)
-			.with_timestamp(1588813835)
-			.add_transaction(transaction.clone())
-			.mine(U256::from(2).pow(254.into())).unwrap();
-
-		let tx_id = transaction.tx_id();
-		let proof = block.merkle_proof(&[tx_id]).unwrap().try_format().unwrap();
-		let raw_tx = transaction.format_with(true);
-
-		let raw_block_header = RawBlockHeader::from_bytes(&block.header.try_format().unwrap()).unwrap();
-		let block_header = BtcRelay::<T>::parse_raw_block_header(&raw_block_header).unwrap();
-
-		BtcRelay::<T>::_store_block_header(&relayer_id, block_header).unwrap();
-		Security::<T>::set_active_block_number(Security::<T>::active_block_number() +
-BtcRelay::<T>::parachain_confirmations() + 1u32.into());
+		let (validators, organizations) = get_validators_and_organizations::<T>();
+		StellarRelay::<T>::_update_tier_1_validator_set(validators, organizations).unwrap();
+		let (tx_env_xdr_encoded, scp_envs_xdr_encoded, tx_set_xdr_encoded) = build_dummy_proof_for::<T>(redeem_id, true);
 
 		assert_ok!(Oracle::<T>::_set_exchange_rate(get_collateral_currency_id::<T>(),
 			UnsignedFixedPoint::<T>::one()
 		));
-	}: _(RawOrigin::Signed(vault_id.account_id.clone()), redeem_id, proof, raw_tx)
+	}: _(RawOrigin::Signed(vault_id.account_id.clone()), redeem_id, tx_env_xdr_encoded, scp_envs_xdr_encoded, tx_set_xdr_encoded)
 
 	cancel_redeem_reimburse {
 		let origin: T::AccountId = account("Origin", 0, 0);
@@ -322,11 +207,9 @@ BtcRelay::<T>::parachain_confirmations() + 1u32.into());
 		let mut redeem_request = test_request::<T>(&vault_id);
 		redeem_request.redeemer = origin.clone();
 		redeem_request.opentime = Security::<T>::active_block_number();
-		redeem_request.btc_height = BtcRelay::<T>::get_best_block_height();
 		Redeem::<T>::insert_redeem_request(&redeem_id, &redeem_request);
 
 		// expire redeem request
-		mine_blocks_until_expiry::<T>(&redeem_request);
 		Security::<T>::set_active_block_number(Security::<T>::active_block_number() + Redeem::<T>::redeem_period() + 100u32.into());
 
 		register_public_key::<T>(vault_id.clone());
@@ -360,7 +243,6 @@ BtcRelay::<T>::parachain_confirmations() + 1u32.into());
 		Redeem::<T>::insert_redeem_request(&redeem_id, &redeem_request);
 
 		// expire redeem request
-		mine_blocks_until_expiry::<T>(&redeem_request);
 		Security::<T>::set_active_block_number(Security::<T>::active_block_number() + Redeem::<T>::redeem_period() + 100u32.into());
 
 		register_public_key::<T>(vault_id.clone());
